+++ pytorch_lightning/core/saving.py
@@ -1,3 +1,9 @@
+import os
+import csv
+import logging as log
+from argparse import Namespace
+
+
 class ModelIO(object):
 
     def on_load_checkpoint(self, checkpoint):
@@ -28,3 +34,34 @@ class ModelIO(object):
         Hook to do whatever you need right before Slurm manager loads the model
         :return:
         """
+
+
+def load_hparams_from_tags_csv(tags_csv):
+    if not os.path.isfile(tags_csv):
+        log.warning(f'Missing Tags: {tags_csv}.')
+        return Namespace()
+
+    tags = {}
+    with open(tags_csv) as f:
+        csv_reader = csv.reader(f, delimiter=',')
+        for row in list(csv_reader)[1:]:
+            tags[row[0]] = convert(row[1])
+    ns = Namespace(**tags)
+    return ns
+
+
+def convert(val):
+    constructors = [int, float, str]
+
+    if isinstance(val, str):
+        if val.lower() == 'true':
+            return True
+        if val.lower() == 'false':
+            return False
+
+    for c in constructors:
+        try:
+            return c(val)
+        except ValueError:
+            pass
+    return val
